GRIM: Split detection features & adapt to new plugins
